// frontend/src/app/openapi/model/fine-grained-resource-output.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration Manager API
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */

import { ProjectUserScopesOutput } from './project-user-scopes-output';
import { AdminScopesOutput } from './admin-scopes-output';
import { UserScopesOutput } from './user-scopes-output';


export interface FineGrainedResourceOutput { 
    user: UserScopesOutput;
    admin: AdminScopesOutput;
    /**
     * Project Slug / Resource mapping.
     */
    projects: { [key: string]: ProjectUserScopesOutput; };
}