// frontend/src/app/openapi/model/user-scopes-output.ts
/*
 * SPDX-FileCopyrightText: Copyright DB InfraGO AG and contributors
 * SPDX-License-Identifier: Apache-2.0
 *
 * Capella Collaboration Manager API
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * Do not edit the class manually.
 + To generate a new version, run `make openapi` in the root directory of this repository.
 */



export interface UserScopesOutput { 
    /**
     * Manage sessions of your own user. With the UPDATE permission a session can be shared with another user and files can be modified via the file browser.
     */
    sessions: Array<UserScopesOutput.SessionsEnum>;
    /**
     * Create new projects
     */
    projects: Array<UserScopesOutput.ProjectsEnum>;
    /**
     * Manage Personal Access Tokens. Use with caution! This permission can be used to escalate privileges by creating tokens with a larger scope.
     */
    tokens: Array<UserScopesOutput.TokensEnum>;
    /**
     * Submit feedback
     */
    feedback: Array<UserScopesOutput.FeedbackEnum>;
}
export namespace UserScopesOutput {
    export const SessionsEnum = {
        Get: 'GET',
        Create: 'CREATE',
        Update: 'UPDATE',
        Delete: 'DELETE'
    } as const;
    export type SessionsEnum = typeof SessionsEnum[keyof typeof SessionsEnum];
    export const ProjectsEnum = {
        Create: 'CREATE'
    } as const;
    export type ProjectsEnum = typeof ProjectsEnum[keyof typeof ProjectsEnum];
    export const TokensEnum = {
        Get: 'GET',
        Create: 'CREATE',
        Delete: 'DELETE'
    } as const;
    export type TokensEnum = typeof TokensEnum[keyof typeof TokensEnum];
    export const FeedbackEnum = {
        Create: 'CREATE'
    } as const;
    export type FeedbackEnum = typeof FeedbackEnum[keyof typeof FeedbackEnum];
}